# web/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rust Tunnel API",
    "description": "RSTunnel 管理面 REST API（§21）",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/nodes": {
      "get": {
        "tags": [
          "nodes"
        ],
        "operationId": "list_nodes",
        "responses": {
          "200": {
            "description": "List nodes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Node"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "nodes"
        ],
        "operationId": "create_node",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Node created (bootstrap token shown once)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateNodeResponse"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate name"
          }
        }
      }
    },
    "/api/v1/nodes/{id}": {
      "get": {
        "tags": [
          "nodes"
        ],
        "operationId": "get_node",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Node id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Node detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Node"
                }
              }
            }
          },
          "404": {
            "description": "Node not found"
          }
        }
      },
      "delete": {
        "tags": [
          "nodes"
        ],
        "operationId": "delete_node",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Node id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Node deleted"
          },
          "404": {
            "description": "Node not found"
          }
        }
      },
      "patch": {
        "tags": [
          "nodes"
        ],
        "operationId": "update_node",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Node id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Node updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Node"
                }
              }
            }
          },
          "404": {
            "description": "Node not found"
          }
        }
      }
    },
    "/api/v1/nodes/{id}/credentials": {
      "post": {
        "tags": [
          "nodes"
        ],
        "summary": "为 node 签发运行时凭据（§5/§71）。明文 token 仅此一次返回。",
        "operationId": "create_credential",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Node id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCredentialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Credential created (token shown once)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialResponse"
                }
              }
            }
          },
          "404": {
            "description": "Node not found"
          }
        }
      }
    },
    "/api/v1/nodes/{id}/credentials/{credential_id}/revoke": {
      "post": {
        "tags": [
          "nodes"
        ],
        "summary": "吊销凭据（§71）：写 `revoked_at`，之后 Agent 无法认证。",
        "operationId": "revoke_credential",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Node id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "credential_id",
            "in": "path",
            "description": "Credential id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Credential revoked"
          },
          "404": {
            "description": "Credential not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateCredentialRequest": {
        "type": "object",
        "properties": {
          "expires_at": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string"
          }
        }
      },
      "CreateNodeRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          }
        }
      },
      "CreateNodeResponse": {
        "type": "object",
        "description": "创建 Node 的响应：`node` + 一次性 bootstrap token（明文仅此一次，§4/§94）。",
        "required": [
          "node",
          "bootstrap_token"
        ],
        "properties": {
          "bootstrap_token": {
            "type": "string"
          },
          "node": {
            "$ref": "#/components/schemas/Node"
          }
        }
      },
...
      "CredentialResponse": {
        "type": "object",
        "description": "新凭据响应：`token` 为明文，仅创建时返回一次（§71）。",
        "required": [
          "id",
          "token"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
...
      "Node": {
        "type": "object",
        "description": "Node 响应对象（api.md §5）。",
        "required": [
          "id",
          "name",
          "status",
          "config_version",
          "applied_config_version",
          "config_status"
        ],
        "properties": {
          "agent_version": {
            "type": "string",
            "nullable": true
          },
          "applied_config_version": {
            "type": "integer",
            "format": "int64"
          },
          "architecture": {
            "type": "string",
            "nullable": true
          },
          "config_status": {
            "type": "string"
          },
          "config_version": {
            "type": "integer",
            "format": "int64"
          },
          "connected_at": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "hostname": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "last_seen_at": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "platform": {
            "type": "string",
            "nullable": true
          },
          "remote_addr": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "UpdateNodeRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  },
...
}